requirements/base.txt
argon2-cffi==18.1.0
redis>=2.10.5
celery==4.2.1 

django==2.1.1
django-model-utils==3.1.2
